+++ R/calculatePD.R
@@ -43,7 +43,7 @@
 #' @export
 
 calculatePD <- function(x, justDF = FALSE, include_root = TRUE) {
-  if (class(x) != "phyloseq") {
+  if (is(x) != "phyloseq") {
     stop("Input is not an object of phyloseq class")
   }