+++ R/checkFilterEffect.R
@@ -56,7 +56,7 @@ checkAbundanceCutOffs <- function(x,
                                   abundance_thres = NULL,
                                   prevalence_thres = NULL){
 
-  if (class(x) != "phyloseq") {
+  if (!is(x, "phyloseq")) {
     stop("Input is not an object of phyloseq class")
   }
   if(is.null(abundance_thres) || is.null(prevalence_thres)){