+++ R/fetchSequence.R
@@ -33,7 +33,7 @@
 fetchSequence <- function(x,
                           seq_id = NULL,
                           as_DNAStringSet = TRUE) {
-  if (class(x) != "phyloseq") {
+  if (!is(x, "phyloseq")) {
     stop("Input is not an object of phyloseq class")
   }