+++ R/getPrevalence.R
@@ -65,7 +65,7 @@ getPrevalence <- function(x,
   # global vars
   Taxa <- NULL
 
-  if (class(x) != "phyloseq") {
+  if (!is(x, "phyloseq")) {
     stop("Input is not an object of phyloseq class")
   }