+++ R/pullPhyloseq.R
@@ -51,7 +51,7 @@ NULL
 #'
 pullSampleAbundance <- function(x, ...) {
 
-  if (class(x) != "phyloseq") {
+  if (!is(x, "phyloseq")) {
     stop("Input is not an object of phyloseq class")
   }
 
@@ -73,7 +73,7 @@ pullSampleAbundance <- function(x, ...) {
 #'
 pullTaxaAbundance <- function(x, ...) {
 
-  if (class(x) != "phyloseq") {
+  if (!is(x, "phyloseq")) {
     stop("Input is not an object of phyloseq class")
   }
 
@@ -99,7 +99,7 @@ pullTaxaAbundance <- function(x, ...) {
 #'
 pullTaxaRank <- function(x, ...) {
 
-  if (class(x) != "phyloseq") {
+  if (!is(x, "phyloseq")) {
     stop("Input is not an object of phyloseq class")
   }