+++ R/separateSampleColumn.R
@@ -36,7 +36,7 @@
 #'
 separateSampleColumn <- function(x, ...){
 
-  if (class(x) != "phyloseq") {
+  if (!is(x, "phyloseq")) {
     stop("Input is not an object of phyloseq class")
   }