+++ R/uniteSampleColumn.R
@@ -35,7 +35,7 @@
 #'
 uniteSampleColumn <- function(x, ...){
 
-  if (class(x) != "phyloseq") {
+  if (!is(x, "phyloseq")) {
     stop("Input is not an object of phyloseq class")
   }